Filled in Column for Web Components based on Mark Hakkinen and Jason Whites email
Here is the email this was based from

1. Works for non-image content

Yes.

Technically, the DIAGRAMMER component could provide a set of alternates
to any markup.

2. Can hold rich content

Yes.

Yes. In our science study, it was an animated simulation.

3. Same technique for different objects

Yes, if I’m understanding correctly. It’s the same for different HTML
elements.

The DIAGRAMMER custom element would be used for any object type
required alternates. What would be differ, potentially, would be which
alternates might be included as children.

4. Reusable in multiple contexts

I don’t understand what this means and how it differs from item 3.

Do you mean that content marked up as a complete image description
could be reused in different contexts?

5. Description clearly associated with object

In the markup, yes, and also to the user if the user interface is
appropriate.

Agree.

6. Clear which associated object is the description

If you mean, “clear to the user”, then this depends on how the user
interface implemented by the Web component is done. It’s certainly
possible.

Agree. We may even leverage the new aria-roledescription to explicitly
label the description type.

7. Not bloat primary content page

I don’t know what the criteria are for this item. There needs to be an
HTML import for the Web component, then the custom element defined by
the Web component has to be used somewhere in the HTML document. The
descriptions/alternatives are included as part of the same page in our
implementation, but it should be possible to provide a URI in an
attribute and then use scripting to retrieve the associated resource.
Whether this constitutes “bloat” (a pejorative term which also happens
to be ambiguous) or not depends on how you define it. I would prefer
avoidance of pejorative terms in such criteria.

Use of URI (e.g. “src” attributes) on individual description elements
would allow “out of document” content, perhaps from a repository.

8. Exposed meaningfully to the accessibility APIs and UA

There isn’t an accessibility API feature for exposing it.

Yes, in that what is available to the accessibility API is what is in
the DOM, with appropriate labeling (again, perhaps made more meaningful
by aria-roledescription.  The default UI would be visible to AT.  If
the author decides to provide their own UI to the alternate content,
then they have to play by the rules for accessibility.

9. Able to pass structured content to an additional specialized user
agent

I think Mark implemented a feature that would invoke a printing API to
enable the user to submit a tactile graphic for embossing. Does this
qualify?

Are you suggesting braille display, or the embosser as suggested by
Jason?

10. Discoverable meaningfully by any current AT.

I don’t know what this means. ARIA does not have a landmark role for
descriptions, so if you’re thinking about navigation then the answer is
“no”. It isn’t clear what this criterion means that would differ from
the item concerned with accessibility APIs above.

By the AT (as a recognized landmark), or by the end-user of the AT?  In
either case, I think the answer would be yes.

11. Skippable meaningfully by any current AT.

If it provides its own user interface (as it does in Mark’s prototype)
then the descriptions can be skipped but the user interface to
hide/reveal them can’t. If one or more of the alternatives are
displayed in response to Web application configuration settings or a
user’s personal needs/preferences profile, then I suppose it depends on
what the enclosing element is and whether the AT can skip it.

For the alternate content, it would only be displayed if requested, and
thus skippable in that sense if display not. Depending upon the markup
used in the actual implementation of the component, I think
skippability of the control structure is not an issue,

12. Hideable from the visual meaningfully by any current AT

I suppose it could all be positioned off-screen or otherwise hidden,
e.g., in response to a configuration setting. This wouldn’t be done by
any AT however; it would be a function of the Web component or the Web
application.

Yes.  We would select the most robust hiding technique.

13. Able to be accessed without using a screen reader in any current UA

Yes.

Yes.

Unfortunately, your criteria are poorly enough defined to make their
value in comparing different solutions dubious. I don’t know how to
answer some of them, and I’m not confident that different people will
answer consistently or that there is a clear meaning/rationale in place
to explain and justify the list.

Please let me know how I can help further. I hope the above comments
are of some assistance.

Regards,

Jason.